+++ ckb-sdk/src/lib.rs
@@ -3,8 +3,8 @@ mod chain;
 mod error;
 mod index;
 mod kvdb;
+mod local;
 mod rpc;
-mod transaction;
 mod util;
 
 pub mod wallet;
@@ -22,7 +22,7 @@ pub use index::{
 pub use kvdb::{KVReader, KVTxn, RocksReader, RocksTxn};
 pub use rpc::HttpRpcClient;
 
-pub use transaction::{
+pub use local::{
     from_local_cell_out_point, to_local_cell_out_point, CellAliasManager, CellInputManager,
     CellManager, ScriptManager, TransactionManager, VerifyResult,
 };
